+++ src/baseConfig.ts
