+++ src/cloudfrontSPAWebsiteStack.ts
