+++ src/cloudfrontStaticWebsiteStack.ts
