+++ src/helper/uploadDirToS3.ts
