+++ src/index.ts
@@ -4,3 +4,5 @@
 // SPDX-License-Identifier: MPL-2.0
 export * from "./cloudfrontStaticWebsiteStack";
 export * from "./cloudfrontSPAWebsiteStack";
+export * from "./baseConfig";
+export * from "./helper";
