+++ tsconfig.json
@@ -17,7 +17,9 @@
     "noImplicitThis": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "outDir": "dist",
     "resolveJsonModule": true,
+    "rootDir": "src",
     "strict": true,
     "strictNullChecks": true,
     "strictPropertyInitialization": true,
